split ebs and dlm class
