+++ package/compute/ebs.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
 
-"""Module deleting all aws ebs volume and dlm policie resources."""
+"""Module deleting all aws ebs volume."""
 
 import logging
 
 import boto3
 
-from botocore.exceptions import ClientError, EndpointConnectionError
+from botocore.exceptions import ClientError
 
 
 class NukeEbs:
@@ -16,22 +16,14 @@ class NukeEbs:
         """Initialize ebs nuke."""
         if region_name:
             self.ec2 = boto3.client("ec2", region_name=region_name)
-            self.dlm = boto3.client("dlm", region_name=region_name)
         else:
             self.ec2 = boto3.client("ec2")
-            self.dlm = boto3.client("dlm")
-
-        try:
-            self.dlm.get_lifecycle_policies()
-        except EndpointConnectionError:
-            print("Dlm resource is not available in this aws region")
-            return
 
     def nuke(self, older_than_seconds):
-        """Ebs and dlm policies deleting function.
+        """Ebs deleting function.
 
-        Deleting all ebs volumes and dlm policy resources with
-        a timestamp greater than older_than_seconds.
+        Deleting all ebs volumes resources with a timestamp
+        greater than older_than_seconds.
 
         :param int older_than_seconds:
             The timestamp in seconds used from which the aws resource
@@ -49,14 +41,6 @@ class NukeEbs:
                     logging.info("volume %s has already been deleted", volume)
                 else:
                     logging.error("Unexpected error: %s", e)
-
-        # Deletes snpashot lifecyle policy
-        for policy in self.list_policy(older_than_seconds):
-            try:
-                self.dlm.delete_lifecycle_policy(PolicyId=policy)
-                print("Nuke EBS Lifecycle Policy {0}".format(policy))
-            except ClientError as e:
-                logging.error("Unexpected error: %s", e)
 
     def list_ebs(self, time_delete):
         """Ebs volume list function.
@@ -76,24 +60,3 @@ class NukeEbs:
             for volume in page["Volumes"]:
                 if volume["CreateTime"].timestamp() < time_delete:
                     yield volume["VolumeId"]
-
-    def list_policy(self, time_delete):
-        """Data Lifecycle Policies list function.
-
-        Returns the IDs of all Data Lifecycle Policies with
-        a timestamp lower than time_delete.
-
-        :param int time_delete:
-            Timestamp in seconds used for filter Data Lifecycle policies
-
-        :yield Iterator[str]:
-            Data Lifecycle policies IDs
-        """
-        response = self.dlm.get_lifecycle_policies()
-
-        for policy in response["Policies"]:
-            detailed = self.dlm.get_lifecycle_policy(
-                PolicyId=policy["PolicyId"]
-            )
-            if detailed["Policy"]["DateCreated"].timestamp() < time_delete:
-                yield policy["PolicyId"]
